@@ -10,34 +10,34 @@
         "ext-iconv": "*",
         "ext-json": "*",
         "composer/package-versions-deprecated": "^1.11",
-        "elao/enum": "2.x-dev",
+        "elao/enum": "^2.0",
         "league/glide-symfony": "^2.0",
         "phpdocumentor/reflection-docblock": "^5.2",
         "sensio/framework-extra-bundle": "^6.1",
         "stenope/stenope": "0.x-dev",
-        "symfony/asset": "5.4.*",
-        "symfony/console": "5.4.*",
-        "symfony/dotenv": "5.4.*",
-        "symfony/expression-language": "5.4.*",
+        "symfony/asset": "^6.1@RC",
+        "symfony/console": "^6.1@RC",
+        "symfony/dotenv": "^6.1@RC",
+        "symfony/expression-language": "^6.1@RC",
         "symfony/flex": "^1.17",
-        "symfony/form": "5.4.*",
-        "symfony/framework-bundle": "5.4.*",
-        "symfony/http-client": "5.4.*",
-        "symfony/intl": "5.4.*",
-        "symfony/mime": "5.4.*",
+        "symfony/form": "^6.1@RC",
+        "symfony/framework-bundle": "^6.1@RC",
+        "symfony/http-client": "^6.1@RC",
+        "symfony/intl": "^6.1@RC",
+        "symfony/mime": "^6.1@RC",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/process": "5.4.*",
-        "symfony/property-access": "5.4.*",
-        "symfony/property-info": "5.4.*",
-        "symfony/runtime": "5.4.*",
-        "symfony/serializer": "5.4.*",
-        "symfony/string": "5.4.*",
-        "symfony/translation": "5.4.*",
-        "symfony/twig-bundle": "5.4.*",
-        "symfony/validator": "5.4.*",
-        "symfony/web-link": "5.4.*",
+        "symfony/process": "^6.1@RC",
+        "symfony/property-access": "^6.1@RC",
+        "symfony/property-info": "^6.1@RC",
+        "symfony/runtime": "^6.1@RC",
+        "symfony/serializer": "^6.1@RC",
+        "symfony/string": "^6.1@RC",
+        "symfony/translation": "^6.1@RC",
+        "symfony/twig-bundle": "^6.1@RC",
+        "symfony/validator": "^6.1@RC",
+        "symfony/web-link": "^6.1@RC",
         "symfony/webpack-encore-bundle": "^1.11",
-        "symfony/yaml": "5.4.*",
+        "symfony/yaml": "^6.1@RC",
         "twig/extra-bundle": "^3.3",
         "twig/intl-extra": "^3.3",
         "twig/string-extra": "^3.3",
@@ -45,19 +45,18 @@
     },
     "require-dev": {
         "ekino/phpstan-banned-code": "^1.0",
-        "elao/accesseo": "^1.0@beta",
         "friendsofphp/php-cs-fixer": "^3.4",
         "phpstan/phpstan": "^1.2",
         "phpstan/phpstan-strict-rules": "^1.1",
         "phpstan/phpstan-symfony": "^1.0",
-        "symfony/browser-kit": "5.4.*",
-        "symfony/css-selector": "5.4.*",
-        "symfony/debug-bundle": "5.4.*",
+        "symfony/browser-kit": "^6.1@RC",
+        "symfony/css-selector": "^6.1@RC",
+        "symfony/debug-bundle": "^6.1@RC",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "5.4.*",
-        "symfony/stopwatch": "5.4.*",
-        "symfony/var-dumper": "5.4.*",
-        "symfony/web-profiler-bundle": "5.4.*"
+        "symfony/phpunit-bridge": "^6.1@RC",
+        "symfony/stopwatch": "^6.1@RC",
+        "symfony/var-dumper": "^6.1@RC",
+        "symfony/web-profiler-bundle": "^6.1@RC"
     },
     "config": {
         "optimize-autoloader": true,
@@ -103,14 +102,14 @@
         ]
     },
     "conflict": {
-        "symfony/symfony": "*",
-        "symfony/dom-crawler": "> 5.4.5"
+        "symfony/symfony": "*"
     },
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.4.*"
+            "require": "6.1.*"
         }
     },
-    "prefer-stable": true
+    "prefer-stable": true,
+    "minimum-stability": "RC"
 }
