@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace App\Command;
 
 use App\Legacy\HtaccessGenerator;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
+#[AsCommand('app:generate-redirections')]
 class GenerateRedirectionsCommand extends Command
 {
-    protected static $defaultName = 'app:generate-redirections';
-
     private HtaccessGenerator $generator;
 
     public function __construct(HtaccessGenerator $generator)
